minute changes

+++ LAB3-Prashast/src/main/java/Armstrong.java
@@ -1,13 +1,3 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-
-/**
- *
- * @author student
- */
-
 //Q1
 
 import java.util.Scanner;
@@ -28,7 +18,7 @@ public class Armstrong {
         while (num != 0){
             a = num%10;
             num = num/10;
-            sum += Math.pow(a,n);
+            sum += (int) Math.pow(a,n);
         }
         if(sum == b)
             System.out.println(b+" is an Armstrong number.");
